sploits/dotFM/requirements.txt
pydub==0.23.1
pypng==0.0.20
Pillow==6.2.0